requirements.txt
discord.py
python-telegram-bot[all]
aiohttp
qrcode
Pillow
python-dotenv